core-foundations/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "core_foundations"
version = "0.1.0"
description = "Core foundational utilities and classes for Forex Trading Platform"
authors = [
    {name = "Forex Trading Platform Team", email = "team@forexplatform.example"}
]
readme = "README.md"
requires-python = ">=3.8,<3.11"
license = {text = "Proprietary"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "fastapi>=0.95.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.22.0,<2.0.0",
    "PyYAML>=6.0,<7.0",
    "pandas-market-calendars>=4.1.4",
    "pytz>=2023.3",
    "requests>=2.28.0,<3.0.0", # Added
    "pymongo>=4.0.0,<5.0.0",   # Added
    "redis>=4.3.0,<5.0.0",     # Added
    "psutil>=5.9.0,<6.0.0",    # Added
    "aiohttp>=3.8.0,<4.0.0",   # Added (optional for health_monitoring)
    "kafka-python>=2.0.0,<3.0.0", # Added (dependency for KafkaEventBus)
    "tenacity>=8.2.3,<9.0.0"    # Added (for retrying logic in event bus)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]
docs = [
    "sphinx>=6.2.1",
    "sphinx-rtd-theme>=1.2.1",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true # Start with ignoring missing imports for external libs
warn_return_any = true
warn_unused_configs = true